refactor: StoreCommonService.existsById 생성
