+++ common/src/main/java/com/food/common/store/business/internal/StoreCommonService.java
@@ -12,4 +12,11 @@ public interface StoreCommonService {
      * @return 주어진 ID를 보유한 가게데이터 또는 찾을 수 없는 경우 Optional#empty()
      */
     Optional<StoreDto> findById(@NotNull Long storeId);
+
+    /**
+     * StoreId로 가게데이터의 존재유무를 리턴한다.
+     * @param storeId 검색할 ID. null일 수 없다.
+     * @return 데이터 존재 유무
+     */
+    boolean existsById(@NotNull Long storeId);
 }
